requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
certifi>=2024.2.2
mcp-server>=0.1.0 